Memperbaiki bug-bug setelah merge pull request
