+++ resources/views/kepsek/submission.blade.php
@@ -1,7 +1,7 @@
 @extends('layouts.layout-submission')
 
 @section('sub-content')
-	<div class="content">
+	<div class="content background-1">
 		<div class="header_submission">
 			<h4>Daftar Pengajuan</h4>
 		</div>
@@ -16,8 +16,9 @@
 		<br><br>
 		<div>
 			<!-- SEMUA ISI TABLE NGE GET DARI DATA PENGAJU -->
-			<table class="table">
+			<table class="table" id="tablesubmission">
 				<tr>
+                    <th>No</th>
 					<th>Pengajuan</th>
 					<th>Pengaju</th>
 					<th>Tanggal</th>
@@ -25,26 +26,32 @@
 					<th>Detail Pengajuan</th>
 				</tr>
 				<tr>
-					<td>Judul Pengajuan</td> <!-- PERLU BACKEND -->
-					<td>Nama Jurusan</td> <!-- PERLU BACKEND -->
-					<td>20-10-20</td> <!-- PERLU BACKEND -->
-					<td>-</td> <!-- PERLU BACKEND -->
-					<td><button type="button" class="btn btn-secondary" data-toggle="modal" data-target="#detail">Lihat Detail</button></td>
-
+                    @foreach ($datasub as $dtsub)
+                    <td>{{ $loop->iteration }}</td> <!-- PERLU BACKEND -->
+					<td>{{$dtsub->judul}}</td> <!-- PERLU BACKEND -->
+					<td>{{$dtsub->nama}}</td> <!-- PERLU BACKEND -->
+                    <td>{{$dtsub->created_at}}</td> <!-- PERLU BACKEND -->
+                    <td>{{$dtsub->status}}</td>
+				<td><a href="" class="btn btn-secondary detail" data-toggle="modal" data-target="#detail-{{$dtsub->id_pengajuan}}">Lihat Detail</a></td>
+                </tr>
+					@endforeach
+					@foreach ($datasub as $data)
+						
+					
 					<!-- Modal -->
-					<div class="modal" id="detail">
+				<div class="modal" id="detail-{{ $data->id_pengajuan }}" tabindex="-1">
 						<div class="modal-dialog">
 							<div class="modal-content">
 								
 								<!-- Header --->
 								<div class="modal-header">
-									<h4 class="modal-title">Judul Pengajuan</h4> <!-- DI GET DARI DATA PENGAJU -->
+									<h4 class="modal-title" id="judul">{{$data->judul}}</h4> <!-- DI GET DARI DATA PENGAJU -->
 									<button type="button" class="close" data-dismiss="modal">&times;</button>
 								</div>
 
 								<!-- Body -->
 								<div class="modal-body top"> <!-- DI GET DARI DATA PENGAJU -->
-									Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+									<p id="deskripsi">{{ $data->deskripsi }}</p>
 
 									<div class="modal-body bottom">
 										<label><b>File Lampiran<b></label>
@@ -54,45 +61,60 @@
 
 								<!-- Footer -->
 								<div class="modal-footer">
-									<button type="button" class="btn btn-success" data-toggle="modal" data-target="#izinkan">Izinkan</button>
-
+								<a class="btn btn-success" data-toggle="modal" data-target="#izinkan-{{$data->id_pengajuan}}">Izinkan</a>
+								<a class="btn btn-danger" data-toggle="modal" data-target="#tidakizinkan-{{$data->id_pengajuan}}">Tidak Izinkan</a>
 									<!-- Modal Izinkan -->
-									<div class="modal fade bs-example-modal-sm" id="izinkan">
+									<div class="modal fade bs-example-modal-sm" id="izinkan-{{$data->id_pengajuan}}" tabindex="1">
 										<div class="modal-dialog modal-sm">
+											<form method="POST" action="/submission/diizinkan">
+												@csrf
 											<div class="modal-content">
 												<div class="modal-header">
-													<h5 class="modal-title">Izinkan</h5>
+													<h5 class="modal-title">Izinkan</h5>	
 													<button type="button" class="close" data-dismiss="modal">&times;</button>
 												</div>
 												<div class="modal-body">
+													<input type="hidden" name="jumlah" value="{{ $data->jumlah }}">
+													<input type="hidden" value="{{ $data->id_pengajuan }}" name="id_pengajuan">
+													<input type="hidden" value="{{ $data->id_transaksi }}" name="id_transaksi">
+													<input type="hidden" value="{{$data->id_dana}}" name="idDana">
+													<input type="hidden" name="status" value="{{$data->status}}">
 													<label class="mid">Berikan Komentar</label>
-													<textarea class="form-control"></textarea>
+													<textarea class="form-control" name="komentar"></textarea>
 												</div>
 												<div class="modal-footer">
 													<input type="submit" name="submit" class="btn btn-primary" value="Submit">
 												</div>
 											</div>
+										</form>
 										</div>
 									</div>
 
-									<button type="button" class="btn btn-danger" data-toggle="modal" data-target="#tidakizinkan">Tidak Izinkan</button>
+									
 
 									<!-- Modal Tidak Izinkan -->
-									<div class="modal fade bs-example-modal-sm" id="tidakizinkan">
+									<div class="modal fade bs-example-modal-sm" id="tidakizinkan-{{$data->id_pengajuan}}" tabindex="1">
 										<div class="modal-dialog modal-sm">
+											<form method="POST" action="/submission/tidakdiizinkan">
+												@csrf
 											<div class="modal-content">
 												<div class="modal-header">
 													<h5 class="modal-title">Tidak Izinkan</h5>
 													<button type="button" class="close" data-dismiss="modal">&times;</button>
 												</div>
 												<div class="modal-body">
+													<input type="hidden" name="jumlah" value="{{ $data->jumlah }}">
+													<input type="hidden" value="{{ $data->id_pengajuan }}" name="id_pengajuan">
+													<input type="hidden" value="{{ $data->id_transaksi }}" name="id_transaksi">
+													<input type="hidden" name="status" value="{{$data->status}}">
 													<label class="mid">Berikan Alasan</label>
-													<textarea class="form-control"></textarea>
+													<textarea class="form-control" name="komentar"></textarea>
 												</div>
 												<div class="modal-footer">
 													<input type="submit" name="submit" class="btn btn-primary" value="Submit">
 												</div>
 											</div>
+											</form>
 										</div>
 									</div>
 
@@ -100,8 +122,9 @@
 							</div>
 						</div>
 					</div>
-				</tr>
+					@endforeach
 			</table>
 		</div>
 	</div>
+	
 @endsection
